refactor: extract scatter-add op into _scatter
Move the indexed-reduce primitive (_scatter_add, _scatter_add_impl)
out of _step.py into a new private mcmcstep/_scatter.py, together with
its per-platform batch-count heuristic (_auto_num_batches,
_final_round), which was a run-time helper misplaced in _state.py. This
gives the core op a self-contained home to grow.
